Move secondary files to tools/ + run black & isort
