+++ fhir2dataset/__init__.py
@@ -1,8 +1,8 @@
-from fhir2dataset.query import Query  # noqa
+import re
+
 from fhir2dataset.fhirrules import FHIRRules  # noqa
 from fhir2dataset.parser import Parser  # noqa
-
-import re
+from fhir2dataset.query import Query  # noqa
 
 
 def sql(sql_query, fhir_api_url=None, token=None):
