+++ fhir2dataset/graphquery.py
@@ -1,12 +1,12 @@
-import networkx as nx
 import logging
-from typing import Type
-from pprint import pformat
 from collections import defaultdict
+from pprint import pformat
+from typing import Type
 
+import networkx as nx
+
+from fhir2dataset.data_class import EdgeInfo, Element, Elements, ResourceAliasInfo, SearchParameter
 from fhir2dataset.fhirrules import FHIRRules
-
-from fhir2dataset.data_class import SearchParameter, ResourceAliasInfo, Element, Elements, EdgeInfo
 
 logger = logging.getLogger(__name__)
 
@@ -15,10 +15,10 @@ class GraphQuery:
     """Class for storing query information in the form of a graph.
 
     Attributes:
-        resources_alias_graph (nx.Graph): The nodes correspond to the aliases involved in
+        resources_graph (nx.Graph): The nodes correspond to the aliases involved in
             the query (filled in the "from") and stops the reference link between 2 aliases
             (filled in the "join").
-        resources_alias_info (dict): Dictionary storing information about each alias:
+        resources_by_alias (dict): Dictionary storing information about each alias:
             * the type of the associated resource
             * the elements that must be retrieved from the json of a resource
             * a boolean indicating whether to return the number of instances of the resource
@@ -39,13 +39,13 @@ class GraphQuery:
         self.fhir_rules = fhir_rules
 
         # to represent the relationships (references) between resources
-        self.resources_alias_graph = nx.Graph()
-        self.resources_alias_info = defaultdict(Type[ResourceAliasInfo])
+        self.resources_graph = nx.Graph()
+        self.resources_by_alias = defaultdict(ResourceAliasInfo)
 
     def build(
         self, select_dict: dict, from_dict: dict, join_dict: dict = None, where_dict: dict = None
     ):
-        """Populates the attributes resources_alias_graph and resources_alias_info according
+        """Populates the attributes resources_graph and resources_by_alias according
         to the information filled in.
 
         Arguments:
@@ -61,18 +61,19 @@ class GraphQuery:
             self._where(**where_dict)
         self._select(**select_dict)
 
-        for resource_alias in self.resources_alias_info.keys():
-            self.resources_alias_info[resource_alias].elements.compute_forest_fhirpaths()
+        # FIXME: Need FHIR2Dataset#96
+        # for resource_alias in self.resources_by_alias.keys():
+        #     self.resources_by_alias[resource_alias].elements.compute_forest_fhirpaths()
 
-        logger.info(f"The nodes are:{self.resources_alias_graph.nodes()}")
+        logger.info(f"The nodes are:{self.resources_graph.nodes()}")
         logger.info("The edges are:")
-        logger.info(pformat(list(self.resources_alias_graph.edges(data=True))))
+        logger.info(pformat(list(self.resources_graph.edges(data=True))))
         logger.info("The information gathered for each node is:")
-        logger.info(pformat(self.resources_alias_info))
+        logger.info(pformat(self.resources_by_alias))
 
     def _from(self, **resource_type_alias):
-        """Initializes the graph nodes contained in resources_alias_graph and the dictionary
-        of resources_alias_info information of the aliases listed in resource_type_alias
+        """Initializes the graph nodes contained in resources_graph and the dictionary
+        of resources_by_alias information of the aliases listed in resource_type_alias
 
         Arguments:
             **resource_type_alias: the key corresponds to the alias and the value to the type of the resource.
@@ -89,17 +90,17 @@ class GraphQuery:
                 ]
             )
 
-            self.resources_alias_graph.add_node(resource_alias)
+            self.resources_graph.add_node(resource_alias)
 
-            self.resources_alias_info[resource_alias] = ResourceAliasInfo(
+            self.resources_by_alias[resource_alias] = ResourceAliasInfo(
                 alias=resource_alias, resource_type=resource_type, elements=elements
             )
 
     def _join(self, **join_as):
         """Builds the reference links between the aliases involved in the query
-        * fills in the elements in attribute resources_alias_info to be retrieved from the
+        * fills in the elements in attribute resources_by_alias to be retrieved from the
           json resource file to be able to make the joins
-        * creates the edges between the nodes of the relevant aliases in the resources_alias_graph
+        * creates the edges between the nodes of the relevant aliases in the resources_graph
           attribute.
 
         Keyword Arguments:
@@ -112,7 +113,7 @@ class GraphQuery:
             join_how = join_how.lower()
             assert join_how in ["inner", "child", "parent", "one"], "Precise how to join"
             for (alias_parent, searchparam_dict) in relationships_dict.items():
-                type_parent = self.resources_alias_info[alias_parent].resource_type
+                type_parent = self.resources_by_alias[alias_parent].resource_type
                 for (searchparam_parent, alias_child) in searchparam_dict.items():
 
                     # Update element to have in table
@@ -124,7 +125,7 @@ class GraphQuery:
                     else:
                         fhirpath_ref = f"({fhirpath}.reference)"
 
-                    self.resources_alias_info[alias_parent].elements.append(
+                    self.resources_by_alias[alias_parent].elements.append(
                         Element(
                             goal="join",
                             col_name=f"join_{searchparam_parent}",
@@ -134,7 +135,7 @@ class GraphQuery:
                     )
 
                     # Update Graph
-                    type_child = self.resources_alias_info[alias_child].resource_type
+                    type_child = self.resources_by_alias[alias_child].resource_type
 
                     searchparam_parent_to_child = f"{searchparam_parent}:{type_child}."
                     # include = f"{type_parent}:{searchparam_parent}:" f"{type_child}"
@@ -155,10 +156,10 @@ class GraphQuery:
                         searchparam_prefix=searchparam_prefix,
                     )
 
-                    self.resources_alias_graph.add_edge(alias_parent, alias_child, info=edge_info)
+                    self.resources_graph.add_edge(alias_parent, alias_child, info=edge_info)
 
     def _where(self, **wheres):
-        """updates the resources_alias_info attribute with the conditions that each alias must meet
+        """updates the resources_by_alias attribute with the conditions that each alias must meet
 
         Keyword Arguments:
             **wheres: the key is an alias, the value is a dictionary containing itself keys which
@@ -184,7 +185,7 @@ class GraphQuery:
                     search_param_obj = SearchParameter(
                         code=search_param, prefix=prefix, value=value
                     )
-                    self.resources_alias_info[resource_alias].elements.append(
+                    self.resources_by_alias[resource_alias].elements.append(
                         Element(
                             goal="where",
                             col_name=f"where_{search_param_obj.code}",
@@ -194,7 +195,7 @@ class GraphQuery:
                     )
 
     def _select(self, **selects):
-        """updates the resources_alias_info attribute with the elements that must be retrieved
+        """updates the resources_by_alias attribute with the elements that must be retrieved
         for each alias
 
         Keyword Arguments:
@@ -204,8 +205,12 @@ class GraphQuery:
         for resource_alias, col_names in selects.items():
             for col_name in col_names:
                 fhirpath = self._check_searchparam_or_fhirpath(resource_alias, col_name)
-                self.resources_alias_info[resource_alias].elements.append(
-                    Element(goal="select", col_name=col_name, fhirpath=fhirpath)
+                self.resources_by_alias[resource_alias].elements.append(
+                    Element(
+                        goal="select",
+                        col_name=col_name,
+                        fhirpath=fhirpath,
+                    )
                 )
 
     def _check_searchparam_or_fhirpath(self, resource_alias: str, searchparam_or_fhirpath: str):
@@ -218,7 +223,7 @@ class GraphQuery:
         Returns:
             str: string of characters corresponding to a fhirpath
         """  # noqa
-        resource_type = self.resources_alias_info[resource_alias].resource_type
+        resource_type = self.resources_by_alias[resource_alias].resource_type
         fhirpath = self.fhir_rules.searchparam_to_fhirpath(
             resource_type=resource_type,
             search_param=searchparam_or_fhirpath,
