+++ fhir2dataset/tools/graph.py
@@ -2,9 +2,9 @@
     Module containing functions useful for the analysis and exploitation of graphs
 """
 
-import networkx as nx
 import logging
 
+import networkx as nx
 
 logger = logging.getLogger(__name__)
